Add support for prerequisite events (#36)
