+++ src/flag.rs
@@ -3,7 +3,6 @@ use serde::Deserialize;
 use crate::eval::{self, Detail, Reason};
 use crate::flag_value::FlagValue;
 use crate::rule::FlagRule;
-use crate::store::Store;
 use crate::user::User;
 use crate::variation::{VariationIndex, VariationOrRollout};
 use crate::BucketResult;
@@ -17,14 +16,14 @@ pub struct Flag {
     #[serde(default)]
     deleted: bool,
 
-    on: bool,
+    pub(crate) on: bool,
 
-    targets: Vec<Target>,
-    rules: Vec<FlagRule>,
-    prerequisites: Vec<Prereq>,
+    pub(crate) targets: Vec<Target>,
+    pub(crate) rules: Vec<FlagRule>,
+    pub(crate) prerequisites: Vec<Prereq>,
 
-    fallthrough: VariationOrRollout,
-    off_variation: Option<VariationIndex>,
+    pub(crate) fallthrough: VariationOrRollout,
+    pub(crate) off_variation: Option<VariationIndex>,
     variations: Vec<FlagValue>,
     pub client_side_availability: ClientSideAvailability,
 
@@ -40,14 +39,14 @@ pub struct Flag {
 
 #[derive(Clone, Debug, Deserialize)]
 pub struct Prereq {
-    key: String,
-    variation: VariationIndex,
+    pub(crate) key: String,
+    pub(crate) variation: VariationIndex,
 }
 
 #[derive(Clone, Debug, Deserialize)]
-struct Target {
-    values: Vec<String>,
-    variation: VariationIndex,
+pub(crate) struct Target {
+    pub(crate) values: Vec<String>,
+    pub(crate) variation: VariationIndex,
 }
 
 #[derive(Clone, Debug, Deserialize)]
@@ -58,68 +57,6 @@ pub struct ClientSideAvailability {
 }
 
 impl Flag {
-    pub fn evaluate(&self, user: &User, store: &dyn Store) -> Detail<&FlagValue> {
-        if !self.on {
-            return self.off_value(Reason::Off);
-        }
-
-        for prereq in &self.prerequisites {
-            if let Some(flag) = store.flag(&prereq.key) {
-                if !flag.on || flag.evaluate(user, store).variation_index != Some(prereq.variation)
-                {
-                    // TODO(ch108583) capture prereq event
-                    return self.off_value(Reason::PrerequisiteFailed {
-                        prerequisite_key: prereq.key.to_string(),
-                    });
-                }
-            } else {
-                return self.off_value(Reason::PrerequisiteFailed {
-                    prerequisite_key: prereq.key.to_string(),
-                });
-            }
-        }
-
-        for target in &self.targets {
-            for value in &target.values {
-                if value == user.key() {
-                    return self.variation(target.variation, Reason::TargetMatch);
-                }
-            }
-        }
-
-        for (rule_index, rule) in self.rules.iter().enumerate() {
-            if rule.matches(&user, store) {
-                let result = self.resolve_variation_or_rollout(&rule.variation_or_rollout, &user);
-                return match result {
-                    Ok(BucketResult {
-                        variation_index,
-                        in_experiment,
-                    }) => {
-                        let reason = Reason::RuleMatch {
-                            rule_index,
-                            rule_id: rule.id.clone(),
-                            in_experiment,
-                        };
-                        self.variation(variation_index, reason)
-                    }
-                    Err(e) => Detail::err(e),
-                };
-            }
-        }
-
-        let result = self.resolve_variation_or_rollout(&self.fallthrough, &user);
-        return match result {
-            Ok(BucketResult {
-                variation_index,
-                in_experiment,
-            }) => {
-                let reason = Reason::Fallthrough { in_experiment };
-                self.variation(variation_index, reason)
-            }
-            Err(e) => Detail::err(e),
-        };
-    }
-
     pub fn is_newer_than(&self, flag: &Flag) -> bool {
         self.version > flag.version
     }
@@ -140,7 +77,7 @@ impl Flag {
         }
     }
 
-    fn resolve_variation_or_rollout(
+    pub(crate) fn resolve_variation_or_rollout(
         &self,
         vr: &VariationOrRollout,
         user: &User,
@@ -189,404 +126,13 @@ impl Flag {
 
 #[cfg(test)]
 mod tests {
-    use std::collections::HashMap;
-
-    use super::*;
-    use maplit::hashmap;
+    use super::Flag;
+    use crate::store::Store;
+    use crate::test_common::TestStore;
+    use crate::{ClientSideAvailability, VariationOrRollout};
     use spectral::prelude::*;
 
     use crate::eval::Reason::*;
-    use crate::flag_value::FlagValue::{Bool, Str};
-    use crate::segment::Segment;
-
-    struct TestStore {
-        flags: HashMap<String, Flag>,
-        segments: HashMap<String, Segment>,
-    }
-
-    impl TestStore {
-        fn new() -> TestStore {
-            TestStore {
-                flags: hashmap! {
-                    "flag".to_string() => serde_json::from_str(r#"{
-                        "key": "flag",
-                        "version": 42,
-                        "on": false,
-                        "targets": [],
-                        "rules": [],
-                        "prerequisites": [],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty"
-                    }"#).unwrap(),
-                    "flagWithRuleExclusion".to_string() => serde_json::from_str(r#"{
-                        "key": "flag",
-                        "version": 42,
-                        "on": false,
-                        "targets": [],
-                        "rules": [
-                            {
-                                "variation": 0,
-                                "id": "6a7755ac-e47a-40ea-9579-a09dd5f061bd",
-                                "clauses": [
-                                    {
-                                        "attribute": "platform",
-                                        "op": "in",
-                                        "values": [
-                                            "web",
-                                            "aem",
-                                            "ios"
-                                        ],
-                                        "negate": false
-                                    }
-                                ],
-                                "trackEvents": true
-                            }
-                        ],
-                        "prerequisites": [],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty",
-                        "trackEvents": false,
-                        "trackEventsFallthrough": true,
-                        "debugEventsUntilDate": 1500000000
-                    }"#).unwrap(),
-                    "flagWithTrackAndDebugEvents".to_string() => serde_json::from_str(r#"{
-                        "key": "flag",
-                        "version": 42,
-                        "on": false,
-                        "targets": [],
-                        "rules": [],
-                        "prerequisites": [],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty",
-                        "trackEvents": true,
-                        "trackEventsFallthrough": true,
-                        "debugEventsUntilDate": 1500000000
-                    }"#).unwrap(),
-                    "flagWithExperiment".to_string() => serde_json::from_str(r#"{
-                        "key": "flagWithExperiment",
-                        "version": 42,
-                        "on": true,
-                        "targets": [],
-                        "rules": [],
-                        "prerequisites": [],
-                        "fallthrough": {
-                          "rollout": {
-                            "kind": "experiment",
-                            "seed": 61,
-                            "variations": [
-                              {"variation": 0, "weight": 10000, "untracked": false},
-                              {"variation": 1, "weight": 20000, "untracked": false},
-                              {"variation": 0, "weight": 70000, "untracked": true}
-                            ]
-                          }
-                        },
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty",
-                        "trackEvents": false,
-                        "trackEventsFallthrough": false,
-                        "debugEventsUntilDate": 1500000000
-                    }"#).unwrap(),
-                    "flagWithRolloutBucketBy".to_string() => serde_json::from_str(r#"{
-                        "key": "rollout",
-                        "on": true,
-                        "prerequisites": [],
-                        "targets": [],
-                        "rules": [
-                            {
-                                "rollout": {
-                                    "variations": [
-                                        {
-                                            "variation": 0,
-                                            "weight": 50000
-                                        },
-                                        {
-                                            "variation": 1,
-                                            "weight": 50000
-                                        },
-                                        {
-                                            "variation": 2,
-                                            "weight": 0
-                                        }
-                                    ],
-                                    "bucketBy": "ld_quid"
-                                },
-                                "id": "6a7755ac-e47a-40ea-9579-a09dd5f061bd",
-                                "clauses": [
-                                    {
-                                        "attribute": "platform",
-                                        "op": "in",
-                                        "values": [
-                                            "web",
-                                            "aem",
-                                            "ios"
-                                        ],
-                                        "negate": false
-                                    }
-                                ],
-                                "trackEvents": false
-                            }
-                        ],
-                        "fallthrough": {
-                            "variation": 2
-                        },
-                        "offVariation": 1,
-                        "variations": [
-                            "rollout1",
-                            "rollout2",
-                            "rollout3"
-                        ],
-                        "clientSideAvailability": {
-                            "usingMobileKey": true,
-                            "usingEnvironmentId": true
-                        },
-                        "clientSide": true,
-                        "salt": "ce2634f116d741a7ad1b7ef363f6f9bc",
-                        "trackEvents": false,
-                        "trackEventsFallthrough": false,
-                        "debugEventsUntilDate": null,
-                        "version": 7,
-                        "deleted": false
-                    }"#).unwrap(),
-                    "flagWithTarget".to_string() => serde_json::from_str(r#"{
-                        "key": "flagWithTarget",
-                        "version": 42,
-                        "on": false,
-                        "targets": [{
-                            "values": ["bob"],
-                            "variation": 0
-                        }],
-                        "rules": [],
-                        "prerequisites": [],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty"
-                    }"#).unwrap(),
-                    "flagWithMissingPrereq".to_string() => serde_json::from_str(r#"{
-                        "key": "flagWithMissingPrereq",
-                        "version": 42,
-                        "on": true,
-                        "targets": [],
-                        "rules": [],
-                        "prerequisites": [{
-                            "key": "badPrereq",
-                            "variation": 1
-                        }],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty"
-                    }"#).unwrap(),
-                    "flagWithOffPrereq".to_string() => serde_json::from_str(r#"{
-                        "key": "flagWithOffPrereq",
-                        "version": 42,
-                        "on": true,
-                        "targets": [],
-                        "rules": [],
-                        "prerequisites": [{
-                            "key": "offPrereq",
-                            "variation": 1
-                        }],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty"
-                    }"#).unwrap(),
-                    "flagWithSatisfiedPrereq".to_string() => serde_json::from_str(r#"{
-                        "key": "flagWithSatisfiedPrereq",
-                        "version": 42,
-                        "on": true,
-                        "targets": [],
-                        "rules": [],
-                        "prerequisites": [{
-                            "key": "prereq",
-                            "variation": 1
-                        }],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty"
-                    }"#).unwrap(),
-                    "prereq".to_string() => serde_json::from_str(r#"{
-                        "key": "prereq",
-                        "version": 42,
-                        "on": true,
-                        "targets": [{
-                            "values": ["bob"],
-                            "variation": 0
-                        }],
-                        "rules": [],
-                        "prerequisites": [],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty"
-                    }"#).unwrap(),
-                    "offPrereq".to_string() => serde_json::from_str(r#"{
-                        "key": "offPrereq",
-                        "version": 42,
-                        "on": false,
-                        "targets": [],
-                        "rules": [],
-                        "prerequisites": [],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 1,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty"
-                    }"#).unwrap(),
-                    "flagWithInRule".to_string() => serde_json::from_str(r#"{
-                        "key": "flagWithInRule",
-                        "version": 42,
-                        "on": false,
-                        "targets": [],
-                        "rules": [{
-                            "id": "in-rule",
-                            "clauses": [{
-                                "attribute": "team",
-                                "negate": false,
-                                "op": "in",
-                                "values": ["Avengers"]
-                            }],
-                            "variation": 0,
-                            "trackEvents": false
-                        }],
-                        "prerequisites": [],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty"
-                    }"#).unwrap(),
-                    "flagWithSegmentMatchRule".to_string() => serde_json::from_str(r#"{
-                        "key": "flagWithSegmentMatchRule",
-                        "version": 42,
-                        "on": true,
-                        "targets": [],
-                        "rules": [{
-                            "id": "match-rule",
-                            "clauses": [{
-                                "attribute": "segmentMatch",
-                                "negate": false,
-                                "op": "segmentMatch",
-                                "values": ["segment"]
-                            }],
-                            "variation": 0,
-                            "trackEvents": false
-                        }],
-                        "prerequisites": [],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty"
-                    }"#).unwrap(),
-                    "flagWithMalformedRule".to_string() => serde_json::from_str(r#"{
-                        "key": "flagWithMalformedRule",
-                        "version": 42,
-                        "on": false,
-                        "targets": [],
-                        "rules": [{
-                            "id": "in-rule",
-                            "clauses": [{
-                                "attribute": "key",
-                                "negate": false,
-                                "op": "in",
-                                "values": ["yes"]
-                            }],
-                            "trackEvents": false
-                        }],
-                        "prerequisites": [],
-                        "fallthrough": {"variation": 1},
-                        "offVariation": 0,
-                        "variations": [false, true],
-                        "clientSideAvailability": {
-                            "usingEnvironmentId": true,
-                            "usingMobileKey": true
-                        },
-                        "salt": "salty"
-                    }"#).unwrap(),
-                },
-                segments: hashmap! {
-                    "segment".to_string() => serde_json::from_str(r#"{
-                        "key": "segment",
-                        "included": ["alice"],
-                        "excluded": [],
-                        "rules": [],
-                        "salt": "salty",
-                        "version": 1
-                    }"#).unwrap()
-                },
-            }
-        }
-
-        fn update_flag(&mut self, flag_key: &str, fun: fn(&mut Flag) -> ()) {
-            let flag = self.flags.get_mut(flag_key).unwrap();
-            fun(flag);
-        }
-    }
-
-    impl Store for TestStore {
-        fn flag(&self, flag_key: &str) -> Option<&Flag> {
-            self.flags.get(flag_key)
-        }
-
-        fn segment(&self, segment_key: &str) -> Option<&Segment> {
-            self.segments.get(segment_key)
-        }
-    }
 
     #[test]
     fn is_experimentation_enabled() {
@@ -642,258 +188,6 @@ mod tests {
                 in_experiment: false,
             }))
             .is_false();
-    }
-
-    #[test]
-    fn test_eval_flag_basic() {
-        let store = TestStore::new();
-        let alice = User::with_key("alice").build(); // not targeted
-        let bob = User::with_key("bob").build(); // targeted
-        let mut flag = store.flag("flagWithTarget").unwrap().clone();
-
-        assert!(!flag.on);
-        let detail = flag.evaluate(&alice, &store);
-        assert_that!(detail.value).contains_value(&Bool(false));
-        assert_that!(detail.variation_index).contains_value(0);
-        assert_that!(detail.reason).is_equal_to(&Off);
-
-        assert_that!(flag.evaluate(&bob, &store)).is_equal_to(&detail);
-
-        // flip off variation
-        flag.off_variation = Some(1);
-        let detail = flag.evaluate(&alice, &store);
-        assert_that!(detail.value).contains_value(&Bool(true));
-        assert_that!(detail.variation_index).contains_value(1);
-
-        // off variation unspecified
-        flag.off_variation = None;
-        let detail = flag.evaluate(&alice, &store);
-        assert_that!(detail.value).is_none();
-        assert_that!(detail.variation_index).is_none();
-        assert_that!(detail.reason).is_equal_to(&Off);
-
-        // flip targeting on
-        flag.on = true;
-        let detail = flag.evaluate(&alice, &store);
-        assert_that!(detail.value).contains_value(&Bool(true));
-        assert_that!(detail.variation_index).contains_value(1);
-        assert_that!(detail.reason).is_equal_to(&Fallthrough {
-            in_experiment: false,
-        });
-
-        let detail = flag.evaluate(&bob, &store);
-        assert_that!(detail.value).contains_value(&Bool(false));
-        assert_that!(detail.variation_index).contains_value(0);
-        assert_that!(detail.reason).is_equal_to(&TargetMatch);
-
-        // flip default variation
-        flag.fallthrough = VariationOrRollout::Variation { variation: 0 };
-        let detail = flag.evaluate(&alice, &store);
-        assert_that!(detail.value).contains_value(&Bool(false));
-        assert_that!(detail.variation_index).contains_value(0);
-
-        // bob's reason should still be TargetMatch even though his value is now the default
-        let detail = flag.evaluate(&bob, &store);
-        assert_that!(detail.value).contains_value(&Bool(false));
-        assert_that!(detail.variation_index).contains_value(0);
-        assert_that!(detail.reason).is_equal_to(&TargetMatch);
-    }
-
-    #[test]
-    fn test_eval_flag_rules() {
-        let store = TestStore::new();
-        let alice = User::with_key("alice").build();
-        let bob = User::with_key("bob")
-            .custom(hashmap! {
-                "team".into() => "Avengers".into(),
-            })
-            .build();
-
-        let mut flag = store.flag("flagWithInRule").unwrap().clone();
-
-        assert!(!flag.on);
-        for user in vec![&alice, &bob] {
-            let detail = flag.evaluate(user, &store);
-            assert_that!(detail.value).contains_value(&Bool(false));
-            assert_that!(detail.variation_index).contains_value(0);
-            assert_that!(detail.reason).is_equal_to(&Off);
-        }
-
-        // flip targeting on
-        flag.on = true;
-        let detail = flag.evaluate(&alice, &store);
-        assert_that!(detail.value).contains_value(&Bool(true));
-        assert_that!(detail.variation_index).contains_value(1);
-        assert_that!(detail.reason).is_equal_to(&Fallthrough {
-            in_experiment: false,
-        });
-
-        let detail = flag.evaluate(&bob, &store);
-        assert_that!(detail.value).contains_value(&Bool(false));
-        assert_that!(detail.variation_index).contains_value(0);
-        assert_that!(detail.reason).is_equal_to(&RuleMatch {
-            rule_id: "in-rule".to_string(),
-            rule_index: 0,
-            in_experiment: false,
-        });
-    }
-
-    #[test]
-    fn test_eval_flag_unsatisfied_prereq() {
-        let store = TestStore::new();
-        let flag = store.flag("flagWithMissingPrereq").unwrap().clone();
-        assert!(flag.on);
-
-        let alice = User::with_key("alice").build();
-        let bob = User::with_key("bob").build();
-
-        for user in vec![&alice, &bob] {
-            let detail = flag.evaluate(user, &store);
-            assert_that!(detail.value).contains_value(&Bool(false));
-            assert_that!(detail.reason).is_equal_to(&PrerequisiteFailed {
-                prerequisite_key: "badPrereq".to_string(),
-            });
-        }
-    }
-
-    #[test]
-    fn test_eval_flag_off_prereq() {
-        let store = TestStore::new();
-        let flag = store.flag("flagWithOffPrereq").unwrap().clone();
-        assert!(flag.on);
-
-        let alice = User::with_key("alice").build();
-
-        let detail = flag.evaluate(&alice, &store);
-        assert_that!(detail.value).contains_value(&Bool(false));
-        assert_that!(detail.reason).is_equal_to(&PrerequisiteFailed {
-            prerequisite_key: "offPrereq".to_string(),
-        });
-    }
-
-    #[test]
-    fn test_eval_flag_satisfied_prereq() {
-        let mut store = TestStore::new();
-        let flag = store.flag("flagWithSatisfiedPrereq").unwrap().clone();
-
-        let alice = User::with_key("alice").build();
-        let bob = User::with_key("bob").build();
-
-        let detail = flag.evaluate(&alice, &store);
-        asserting!("alice should pass prereq and see fallthrough")
-            .that(&detail.value)
-            .contains_value(&Bool(true));
-        let detail = flag.evaluate(&bob, &store);
-        asserting!("bob should see prereq failed due to target")
-            .that(&detail.value)
-            .contains_value(&Bool(false));
-        assert_that!(detail.reason).is_equal_to(Reason::PrerequisiteFailed {
-            prerequisite_key: "prereq".to_string(),
-        });
-
-        // prerequisite off
-        store.update_flag("prereq", |flag| (flag.on = false));
-        for user in vec![&alice, &bob] {
-            let detail = flag.evaluate(user, &store);
-            assert_that!(detail.value).contains_value(&Bool(false));
-            assert_that!(detail.reason).is_equal_to(&PrerequisiteFailed {
-                prerequisite_key: "prereq".to_string(),
-            });
-        }
-    }
-
-    #[test]
-    fn test_eval_flag_segments() {
-        let store = TestStore::new();
-        let flag = store.flag("flagWithSegmentMatchRule").unwrap();
-
-        let alice = User::with_key("alice").build();
-        let bob = User::with_key("bob").build();
-
-        let detail = flag.evaluate(&alice, &store);
-        asserting!("alice is in segment, should see false with RuleMatch")
-            .that(&detail.value)
-            .contains_value(&Bool(false));
-        assert_that!(detail.reason).is_equal_to(Reason::RuleMatch {
-            rule_id: "match-rule".to_string(),
-            rule_index: 0,
-            in_experiment: false,
-        });
-        let detail = flag.evaluate(&bob, &store);
-        asserting!("bob is not in segment and should see fallthrough")
-            .that(&detail.value)
-            .contains_value(&Bool(true));
-        assert_that!(detail.reason).is_equal_to(Reason::Fallthrough {
-            in_experiment: false,
-        });
-    }
-
-    #[test]
-    fn test_rollout_flag() {
-        let store = TestStore::new();
-        let flag = store.flag("flagWithRolloutBucketBy").unwrap();
-
-        let alice = User::with_key("anonymous")
-            .custom(hashmap! {
-                "platform".into() => "aem".into(),
-                "ld_quid".into() => "d4ad12cb-392b-4fce-b214-843ad625d6f8".into()
-            })
-            .build();
-
-        let detail = flag.evaluate(&alice, &store);
-        assert_that!(detail.value).contains_value(&Str("rollout1".to_string()));
-    }
-
-    #[test]
-    fn test_experiment_flag() {
-        let store = TestStore::new();
-        let flag = store.flag("flagWithExperiment").unwrap();
-
-        let user_a = User::with_key("userKeyA").build();
-        let detail = flag.evaluate(&user_a, &store);
-        assert_that!(detail.value).contains_value(&Bool(false));
-        assert!(detail.reason.is_in_experiment());
-
-        let user_b = User::with_key("userKeyB").build();
-        let detail = flag.evaluate(&user_b, &store);
-        assert_that!(detail.value).contains_value(&Bool(true));
-        assert!(detail.reason.is_in_experiment());
-
-        let user_c = User::with_key("userKeyC").build();
-        let detail = flag.evaluate(&user_c, &store);
-        assert_that!(detail.value).contains_value(&Bool(false));
-        assert!(!detail.reason.is_in_experiment());
-    }
-
-    #[test]
-    fn test_malformed_rule() {
-        let store = TestStore::new();
-        let mut flag = store.flag("flagWithMalformedRule").unwrap().clone();
-
-        let user_a = User::with_key("no").build();
-        let user_b = User::with_key("yes").build();
-
-        let detail = flag.evaluate(&user_a, &store);
-        assert_that!(detail.value).contains_value(&Bool(false));
-        assert_that!(detail.reason).is_equal_to(Reason::Off);
-
-        let detail = flag.evaluate(&user_b, &store);
-        assert_that!(detail.value).contains_value(&Bool(false));
-        assert_that!(detail.reason).is_equal_to(Reason::Off);
-
-        flag.on = true;
-
-        let detail = flag.evaluate(&user_a, &store);
-        assert_that!(detail.value).contains_value(&Bool(true));
-        assert_that!(detail.reason).is_equal_to(Reason::Fallthrough {
-            in_experiment: false,
-        });
-
-        let detail = flag.evaluate(&user_b, &store);
-        assert_that!(detail.value).is_none();
-        assert_that!(detail.reason).is_equal_to(Reason::Error {
-            error: eval::Error::MalformedFlag,
-        });
     }
 
     #[test]
